Change a i32 to i64 because it's converted from usize.

+++ crates/lgn-editor-srv/src/resource_browser_plugin.rs
@@ -382,7 +382,7 @@ impl Api for Server {
 
         let next_search_token = NextSearchToken {
             next_search_token: "".to_string(),
-            total: u32::try_from(descriptors.len()).unwrap(),
+            total: u64::try_from(descriptors.len()).unwrap(),
             resource_description: descriptors,
         };
